Moved MockBoard to telemetrix.py
